Move `StAnnTx` type family and `txStAnnTxG` to `EraStAnnTx` class
