+++ eras/conway/impl/src/Cardano/Ledger/Conway/Tx.hs
@@ -21,7 +21,6 @@ module Cardano.Ledger.Conway.Tx (
 import Cardano.Ledger.Allegra.Tx (validateTimelock)
 import Cardano.Ledger.Alonzo.Core
 import Cardano.Ledger.Alonzo.Tx (
-  AlonzoStAnnTx (..),
   alonzoMinFeeTx,
   alonzoTxEqRaw,
   auxDataAlonzoTxL,
@@ -51,7 +50,7 @@ import Data.Typeable (Typeable)
 import Data.Word (Word32)
 import GHC.Generics (Generic)
 import GHC.Stack
-import Lens.Micro (Lens', lens, to, (^.))
+import Lens.Micro (Lens', lens, (^.))
 import NoThunks.Class (NoThunks)
 
 instance HasEraTxLevel Tx ConwayEra where
@@ -61,10 +60,6 @@ instance EraTx ConwayEra where
   newtype Tx l ConwayEra = MkConwayTx {unConwayTx :: AlonzoTx l ConwayEra}
     deriving newtype (Eq, Show, NFData, NoThunks, ToCBOR, EncCBOR)
     deriving (Generic)
-
-  type StAnnTx l ConwayEra = AlonzoStAnnTx l ConwayEra
-
-  txStAnnTxG = to $ \AlonzoStAnnTx {asatTx} -> asatTx
 
   mkBasicTx = MkConwayTx . mkBasicAlonzoTx
