+++ eras/conway/impl/src/Cardano/Ledger/Conway/UTxO.hs
@@ -25,7 +25,6 @@ import Cardano.Ledger.Alonzo.UTxO (
   getMintingScriptsNeeded,
   getRewardingScriptsNeeded,
   getSpendingScriptsNeeded,
-  scriptsProvidedAlonzoStAnnTx,
   zipAsIxItem,
  )
 import Cardano.Ledger.Babbage.UTxO (
@@ -151,8 +150,6 @@ instance AlonzoEraUTxO ConwayEra where
   getSupplementalDataHashes = getBabbageSupplementalDataHashes
 
   getSpendingDatum = getBabbageSpendingDatum
-
-  scriptsProvidedStAnnTx = scriptsProvidedAlonzoStAnnTx
 
 getConwayMinFeeTxUtxo ::
   ( EraTx era
