+++ libs/cardano-ledger-core/src/Cardano/Ledger/Core.hs
@@ -162,21 +162,6 @@ class
   where
   data Tx (l :: TxLevel) era
 
-  -- | This is a `Tx` that is annotated with some pre-computed data derived from the ledger state,
-  -- which can be used to avoid redundant computation when a transaction is validated multiple
-  -- times.
-  --
-  -- It is critical to only store here information that satisfies the following property: if the
-  -- ledger state changes in a way that makes the annotated value stale, then some other predicate
-  -- check in the STS rules must independently cause the transaction to be rejected.  Stale
-  -- annotations must never lead to a transaction being silently accepted.
-  --
-  -- For example, if a reference input gets spent, then there must a predicate check that fails on
-  -- missing output, regardless if data from reference inputs is still present in the `StAnnTx`
-  type StAnnTx (l :: TxLevel) era = (r :: Type) | r -> l era
-
-  txStAnnTxG :: SimpleGetter (StAnnTx l era) (Tx l era)
-
   mkBasicTx :: TxBody l era -> Tx l era
 
   bodyTxL :: Lens' (Tx l era) (TxBody l era)
